requirements.txt
annotated-types==0.6.0
arrow==1.3.0
asgiref==3.7.2
blessed==1.20.0
Django==4.2.6
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap5==23.3
django-cors-headers==4.3.0
django-datatable==0.3.1
django-picklefield==3.1
django-q==1.3.9
django-simple-history==3.4.0
et-xmlfile==1.1.0
gunicorn==20.1.0
openpyxl==3.1.2
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.7
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
wcwidth==0.2.8
